Put api calls into module, import into main.js
